+++ crates/miden-testing/tests/auth/mod.rs
@@ -1,3 +1,4 @@
+mod singlesig;
 mod singlesig_acl;
 
 mod multisig;